apps/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-srtm>=0.6.0
numpy>=1.24.0
pillow>=10.0.0
pydantic>=2.0.0
numba>=0.57.0
git+https://github.com/tedaks/pyitm.git
python-dotenv>=1.0.0
rasterio>=1.3.0